Implement absolute paths so code is more readable
