+++ src/pages/ShipOrder/index.tsx
@@ -3,9 +3,9 @@ import Container from "@mui/material/Container";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Paper from "@mui/material/Paper";
-import { Order } from "../../Types";
+import { Order } from "Types";
 import CollapsibleTable from "./OrderList";
-import { getOrders } from "../../storage/readAndWriteOrders";
+import { getOrders } from "storage/readAndWriteOrders";
 import AddIcon from "@mui/icons-material/Add";
 import { useNavigate, Outlet } from "react-router-dom";
