+++ tsconfig.common.json
@@ -14,6 +14,5 @@
     "jsx": "react",
     "checkJs": false
   },
-  "exclude": ["**/node_modules"],
-  "include": [".eslintrc.js", "./src/**/*"]
+  "exclude": ["**/node_modules"]
 }
